#1-final model implementation
