+++ .pylintrc
@@ -17,3 +17,4 @@ max-line-length=120
 
 [DESIGN]
 max-args=10
+min-public-methods=1
